% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-data.R
\name{aggregate_hourly_precipitation}
\alias{aggregate_hourly_precipitation}
\title{Aggregate hourly data to daily precipitation statistics}
\usage{
aggregate_hourly_precipitation(
  df_hourly,
  time_col = "datetime",
  precip_col = "precipitation",
  tz = "UTC"
)
}
\arguments{
\item{df_hourly}{Data frame with hourly precipitation data}

\item{time_col}{Name of the datetime column (must be POSIXct)}

\item{precip_col}{Name of the hourly precipitation column}

\item{tz}{Timezone (default: "UTC")}
}
\value{
A tibble with columns: date, prcp
}
\description{
Aggregate hourly data to daily precipitation statistics
}
