% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-data.R
\name{aggregate_hourly_temperature}
\alias{aggregate_hourly_temperature}
\title{Aggregate hourly data to daily temperature statistics}
\usage{
aggregate_hourly_temperature(
  df_hourly,
  time_col = "datetime",
  temp_col = "temperature",
  tz = "UTC"
)
}
\arguments{
\item{df_hourly}{Data frame with hourly temperature data}

\item{time_col}{Name of the datetime column (must be POSIXct)}

\item{temp_col}{Name of the hourly temperature column}

\item{tz}{Timezone (default: "UTC")}
}
\value{
A tibble with columns: date, tmax, tmin
}
\description{
Aggregate hourly data to daily temperature statistics
}
