% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_indices.R
\name{list_indices}
\alias{list_indices}
\title{List available climate indices}
\usage{
list_indices()
}
\value{
A data frame with available indices and descriptions
}
\description{
List available climate indices
}
